+++ ChangeLog.md
@@ -3,5 +3,6 @@
 ## 1.0
 
 ### Unreleased
+- NEW : automated contract tariff revision feature that updates contracts line prices on a scheduled date and notifies relevant users of the change  - 1.1.0
 - NEW : Create clichaumeil_chaumeilrfa object : - Initial version *19/09/2025* - 1.0.0
 - NEW : Create/Clean clichaumeil module : - Initial version *09/09/2025* - 1.0.0